# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz Platform API",
        "description": "API documentation for Quiz Platform mobile app",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "description": "List all active categories with question counts.",
                "operationId": "ac9c4d978ea4e5b72915ca9728e5b8b1",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Accept-Language"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of categories"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "description": "Get a single category with questions.",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category details"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "parameters": {
            "Accept-Language": {
                "name": "Accept-Language",
                "in": "header",
                "description": "Language code (ar, en)",
                "required": false,
                "schema": {
                    "type": "string",
                    "default": "ar"
                }
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "description": "Enter your bearer token",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}